\name{getBIC-methods}
\docType{methods}
\alias{getBIC-methods}
\alias{getBIC,CClustRes-method}
\alias{getBIC}
\title{get the BIC for choosing the number of clusters}
\description{get the BIC from Curve Clustering results for choosing the number of clusters}
\usage{getBIC(.Object,CCD)}
\arguments{
\item{.Object}{an object of the CClustRes}
\item{CCD}{an object of the CClustData class}
}
\keyword{methods}

