\name{getICL-methods}
\docType{methods}
\alias{getICL-methods}
\alias{getICL,CClustRes-method}
\alias{getICL}
\title{get the ICL criterion for choosing the number of clusters}
\description{get the Integrated Classification Likelihood criterion from Curve Clustering results for choosing the number of clusters}
\usage{getICL(.Object,CCD)}
\arguments{
\item{.Object}{an object of the CClustRes}
\item{CCD}{an object of the CClustData class}
}
\keyword{methods}

